refactor ProgressStatus scope
