+++ assets/src/main/kotlin/org/wycliffeassociates/otter/assets/initialization/InitializeApp.kt
@@ -22,7 +22,7 @@ import io.reactivex.Observable
 import io.reactivex.schedulers.Schedulers
 import io.reactivex.subjects.PublishSubject
 import org.slf4j.LoggerFactory
-import org.wycliffeassociates.otter.common.persistence.config.ProgressStatus
+import org.wycliffeassociates.otter.common.data.ProgressStatus
 import javax.inject.Inject
 
 class InitializeApp @Inject constructor(
