+++ assets/src/main/kotlin/org/wycliffeassociates/otter/assets/initialization/InitializeArtwork.kt
@@ -26,7 +26,7 @@ import org.wycliffeassociates.otter.common.persistence.config.Installable
 import org.wycliffeassociates.otter.common.persistence.repositories.IInstalledEntityRepository
 import javax.inject.Inject
 import org.wycliffeassociates.otter.common.persistence.IDirectoryProvider
-import org.wycliffeassociates.otter.common.persistence.config.ProgressStatus
+import org.wycliffeassociates.otter.common.data.ProgressStatus
 
 class InitializeArtwork @Inject constructor(
     private val installedEntityRepo: IInstalledEntityRepository,
