+++ assets/src/main/kotlin/org/wycliffeassociates/otter/assets/initialization/InitializeUlb.kt
@@ -26,7 +26,7 @@ import org.wycliffeassociates.otter.common.domain.resourcecontainer.ImportExcept
 import org.wycliffeassociates.otter.common.domain.resourcecontainer.ImportResult
 import org.wycliffeassociates.otter.common.persistence.IDirectoryProvider
 import org.wycliffeassociates.otter.common.persistence.config.Installable
-import org.wycliffeassociates.otter.common.persistence.config.ProgressStatus
+import org.wycliffeassociates.otter.common.data.ProgressStatus
 import org.wycliffeassociates.otter.common.persistence.repositories.IInstalledEntityRepository
 import java.io.File
 import javax.inject.Inject
