+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/persistence/config/Initializable.kt
@@ -20,6 +20,7 @@ package org.wycliffeassociates.otter.common.persistence.config
 
 import io.reactivex.Completable
 import io.reactivex.Observer
+import org.wycliffeassociates.otter.common.data.ProgressStatus
 
 /**
  * An initialization task that can be executed.
@@ -30,10 +31,3 @@ interface Initializable {
      * */
     fun exec(progressEmitter: Observer<ProgressStatus>): Completable
 }
-
-data class ProgressStatus(
-    val titleKey: String? = null,
-    val titleMessage: String? = null,
-    val subTitleKey: String? = null,
-    val subTitleMessage: String? = null
-)
