+++ jvm/workbookapp/src/integration-test/kotlin/integrationtest/initialization/TestInitializeSources.kt
@@ -9,7 +9,7 @@ import org.junit.Test
 import org.wycliffeassociates.otter.assets.initialization.InitializeSources
 import org.wycliffeassociates.otter.common.domain.languages.ImportLanguages
 import org.wycliffeassociates.otter.common.persistence.IDirectoryProvider
-import org.wycliffeassociates.otter.common.persistence.config.ProgressStatus
+import org.wycliffeassociates.otter.common.data.ProgressStatus
 import org.wycliffeassociates.otter.common.persistence.repositories.IResourceMetadataRepository
 import org.wycliffeassociates.otter.jvm.workbookapp.persistence.database.AppDatabase
 import java.io.File
